refactor: merge version list test data into test file
